+++ __tests__/spot/sub_account/subAccountFuturesAssetTransferHistory.test.js
@@ -26,7 +26,7 @@ describe('#subAccountFuturesAssetTransferHistory', () => {
 
   it('should get sub account futures asset transfer history', () => {
     const parameters = {
-      email: email,
+      email,
       futuresType,
       recvWindow
     }