+++ src/index.js
@@ -1,1 +1,3 @@
+'use strict'
+
 module.exports.Spot = require('./spot')